all auth endpoint is ok

+++ app/Http/Controllers/Api/AuthController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\User\LoginRequest;
 use App\Http\Requests\User\RegisterRequest;
 use App\Http\Requests\User\VerifyPhoneRequest;
 use App\Http\Controllers\Controller;
+use Illuminate\Validation\ValidationException;
 
 class AuthController extends Controller
 {
@@ -19,46 +20,61 @@ class AuthController extends Controller
 
     public function register(RegisterRequest $request)
     {
-        $data = $request->validated();
+        try {
+            $data = $request->validated();
+            
+            $result = $this->authService->register($data);
 
-        $result = $this->authService->register($data);
+            return response()->json($result, 201);
+        } catch (\Exception $e) {
 
-        return response()->json($result, 201);
+            return response()->json(['message' => $e->getMessage()], 500);
+        }
     }
 
     public function login(LoginRequest $request)
     {
-        $data = $request->validated();
-      
+        try {
+            $data = $request->validated();
+            $result = $this->authService->login($data);
+            return response()->json($result, 201);
 
-        $result = $this->authService->login($data);
-
-        if (!$result) {
-            return response()->json(['message' => 'Invalid credentials or unverified account'], 401);
+        }  catch (\Exception $e) {
+            return response()->json(['message' => $e->getMessage()], 500);
         }
-
-        return response()->json($result);
     }
+    
 
     public function verify(VerifyPhoneRequest $request)
     {
-        $data = $request->validated();
+        try {
+            $data = $request->validated();
+            $isVerified = $this->authService->verifyCode($data['phone'], $data['verification_code']);
 
-        $isVerified = $this->authService->verifyCode($data['phone'], $data['verification_code']);
+            if ($isVerified) {
+                return response()->json(['message' => 'Account verified successfully']);
+            }
 
-        if ($isVerified) {
-            return response()->json(['message' => 'Account verified successfully']);
+            return response()->json(['message' => 'Invalid verification code'], 400);
+        } catch (\Exception $e) {
+            return response()->json(['message' => $e->getMessage()], 500);
         }
-
-        return response()->json(['message' => 'Invalid verification code'], 400);
     }
 
-
     public function logout(Request $request)
-{
-    // Revoke the token that was used to authenticate the current request
-    $request->user()->currentAccessToken()->delete();
-
-    return response()->json(['message' => 'Successfully logged out']);
-}
+    {
+        if (!$request->user()) {
+            return response()->json([
+                'status' => 'fail',
+                'message' => 'Auth token not found. Log in now!',
+            ], 401);
+        }
+        // dd($request->user());
+        try {
+            $request->user()->currentAccessToken()->delete();
+            return response()->json(['message' => 'Successfully logged out'],200);
+        } catch (\Exception $e) {
+            return response()->json(['message' => $e->getMessage()], 500);
+        }
+    }
 }
